Add read button
